refactor: endpoints.ts 디렉토리 변경 -> commons
