+++ client/src/libs/axios.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import endpoints from './endpoints';
+import endpoints from '@/commons/endpoints';
 
 const instance = axios.create({
   baseURL: endpoints.API_BASE_URL,
